requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.5
pydantic-settings==2.9.1

# 데이터베이스
pymysql==1.1.1
sqlalchemy==2.0.41
aiomysql==0.2.0
cryptography==42.0.5

# 환경 설정
python-dotenv==1.1.0

# ML 라이브러리 (최신 버전)
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
transformers==4.52.4
safetensors==0.5.3
sentencepiece==0.2.0
Pillow==11.2.1
numpy==2.2.6

# 웹 스크래핑
requests==2.32.4
beautifulsoup4==4.13.4
duckduckgo-search==4.1.1
googlesearch-python==1.2.3

# FAISS 벡터 검색
faiss-cpu==1.11.0

# 이미지 처리
opencv-python==4.11.0.86

# URL 처리
tldextract==5.3.0

# 사용자 인증
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0

# 추가 유틸리티
aiofiles==23.2.1
python-multipart==0.0.6 